Adds support for foreign key relationships
Whereas before the m2m was supposed to provide basic relationships
between tables, this does it properly by implementing a simple way to
declare foreign key relationships between tables.

This moves the QuerystringDBHelper to a wrapper class, so arbitrary
types of DBHelpers can support querystrings.

This adds two new types of annotations: @DBSortOrder and @UriPath, which
both get attached to the ContentItem. The former sets the default sort
order and the later sets the path (at least according to some
functions.)
(I'm unconvinced that attaching @UriPath to the ContentItem is the best
place for it, but it works for the moment)