Finishes first pass at SearchDBHelper
This adds in a DBHelper that implements Android's SearchManager
interface. A search URI can easily be constructed using this DBHelper
which can have one or more tables registered to it, searching one or
more columns from within that table. Results are presented with
appropriate data URIs so that results can be acted on.

At the moment, this does a somewhat naive LIKE "%query%" search instead
of using the fts module.