Reworks the MIME type system: may break existing code
The MIME type system now uses the table name instead of the path name.
Previously, this would create issues when there was one type that had
two paths: each would have a distinct MIME type, even though the content
was the same.

Adds unit tests for MIME types to ensure that the defaults are working
correctly.