Adds an official parser for QuerystringWrapper; changes syntax
In order to support more complex queries, this adds a full-fledged
parser to parse URL options. The parser is generated with bison and both
the generated code, as well as the source are included in this
repository in order to minimize dependencies.

As a result of the possibility for more complex queries, the syntax has
been updated. This will be updated in the documentation in further
revisions.