Allow implementors to control network syncing
Allow implementors to control the network syncing notification by
analyzing the ContentValues that are used to update or insert the
record. Deletes are always treated as not requiring a sync as by the
time the delete is completed the sync adapter will find it hard to sync
now deleted data.